requirements.txt
alembic==1.0.11
amqp==2.5.1
billiard==3.6.1.0
blinker==1.4
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
gevent==1.4.0
greenlet==0.4.15
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.4
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
Pillow==6.2.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
redis==3.3.8
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
ujson==1.35
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
zipp==0.6.0
